rename vr-stereo-toggle files to vr-stereo
