--- src/components/vr-stereo-toggle/index.js
@@ -1,1 +1,0 @@
-export {VrStereo} from './vr-stereo-toggle';
