+++ src/components/vr-stereo/_vr-stereo.scss
