+++ src/components/vr-stereo/index.js
@@ -1,0 +1,1 @@
+export {VrStereo} from './vr-stereo';
