+++ src/styles/style.scss
@@ -35,7 +35,7 @@
 @import '../components/fullscreen/fullscreen';
 @import '../components/watermark/watermark';
 @import '../components/language/language';
-@import '../components/vr-stereo-toggle/vr-stereo-toggle';
+@import '../components/vr-stereo/vr-stereo';
 @import '../components/cast/cast';
 @import '../components/cast-overlay/cast-overlay';
 @import '../components/cast-on-tv/cast-on-tv';
